+++ WillyNet.Charlas.Core.Application/Features/Users/Queries/GetAllUsers/GetAllUsersQuery.cs
@@ -18,7 +18,7 @@ namespace WillyNet.Charlas.Core.Application.Features.Users.Queries.GetAllUsers
     {        
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public short Dni { get; set; }
+        public int Dni { get; set; }
         public int PageNumber { get; set; }
         public int PageSize { get; set; }
     }
